+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 # rollup-plugin-css-chunks changelog
 
+## Development
+
+* fix: no-treeshake causing inclusion of dead CSS and "phantom" js modules (#7);
+
 ## 1.2.9 (30/11/2020)
 
 * build: add type definitions.